Remove type only file and move content to introspect_handler file
